Renamed "back_end.rs" to "graphics.rs"
Closes https://github.com/PistonDevelopers/graphics/issues/849
